Fixes for the previous fixes.

+++ common/math/vector.h
@@ -460,8 +460,8 @@ inline const EAngle VectorAngles( const Vector& vecForward )
 {
 	EAngle angReturn(0, 0, 0);
 
-	angReturn.p = atan2(vecForward.z, sqrt(vecForward.x*vecForward.x + vecForward.y*vecForward.y)) * 180/M_PI;
-	angReturn.y = atan2(vecForward.y, vecForward.x) * 180/M_PI;
+	angReturn.p = atan2(vecForward.z, sqrt(vecForward.x*vecForward.x + vecForward.y*vecForward.y)) * (float)(180/M_PI);
+	angReturn.y = atan2(vecForward.y, vecForward.x) * (float)(180/M_PI);
 
 	return angReturn;
 }
